Conflicts:
	CC/TrdKCluster.C

+++ CC/status.C
@@ -1,4 +1,4 @@
-//  $Id: status.C,v 1.60 2013/12/20 22:21:04 choutko Exp $
+//  $Id$
 // Author V.Choutko.
 #include "status.h"
 #include "snode.h"
@@ -318,7 +318,7 @@ integer AMSStatus::_statusok(statusI status){
         else {
           uinteger st=STATUSFFKEY.status[i];
           uinteger nbit=32.*log(2.)/log(number(STATUSFFKEY.status[nsta+1]))+0.5;
-          for (int j=0;j<nbit;j++){
+          for (unsigned int j=0;j<nbit;j++){
             uinteger stbit=(st%STATUSFFKEY.status[nsta+1])>0?1:0;
             if((stbit<<j) & (1<<Status[i])){
           //    if(i==15){
@@ -360,7 +360,7 @@ integer AMSStatus::getnextok(){
 const int maxthr=256;
 const long long max32=4294967296LL;
   uint64 off64=_Offset+_Status[i].off;
-if(i<maxthr && _Status[i].off>_Status[maxthr].off && off64>=max32){
+if(i<maxthr && _Status[i].off>_Status[maxthr].off && off64>=uint64(max32)){
    cerr<<"AMSSetatus::getnextok-W-32bitProblemfound "<<i<<" "<<_Status[i].off<<" "<<_Offset<<" "<<off64%max32<<endl;
      off64=off64%max32;
 }
